test(vcs): dedicated unit tests for gitExec, vcsBootstrap; cron edge case
