+++ packages/service/src/vcs/VcsManager.test.ts
@@ -1,10 +1,10 @@
 /**
  * @module vcs/VcsManager.test
- * Tests for VcsManager static and instance methods.
+ * Tests for VcsManager instance methods.
  */
 
 import { execFile } from 'node:child_process';
-import { access, mkdtemp, readFile, rm, writeFile } from 'node:fs/promises';
+import { mkdtemp, rm, writeFile } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { promisify } from 'node:util';
@@ -14,7 +14,7 @@ import pino from 'pino';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { CommitMessageGenerator } from './CommitMessageGenerator';
-import { checkGitAvailable, ensureGitignore, initRepo } from './vcsBootstrap';
+import { initRepo } from './vcsBootstrap';
 import { VcsManager } from './VcsManager';
 
 const execFileAsync = promisify(execFile);
@@ -45,99 +45,6 @@ async function commitCount(cwd: string): Promise<number> {
     return 0;
   }
 }
-
-// ─── Static methods ───
-
-describe('checkGitAvailable', () => {
-  it('returns true when git is available', async () => {
-    const result = await checkGitAvailable();
-    expect(result).toBe(true);
-  });
-});
-
-describe('initRepo', () => {
-  let tempDir: string;
-
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'vcs-init-'));
-  });
-
-  it('initializes a git repo in an empty directory', async () => {
-    await initRepo(tempDir);
-    await expect(access(join(tempDir, '.git'))).resolves.toBeUndefined();
-  });
-
-  it('is idempotent — does not fail on existing repo', async () => {
-    await initRepo(tempDir);
-    await initRepo(tempDir);
-    await expect(access(join(tempDir, '.git'))).resolves.toBeUndefined();
-  });
-});
-
-describe('ensureGitignore', () => {
-  let tempDir: string;
-
-  beforeEach(async () => {
-    tempDir = await mkdtemp(join(tmpdir(), 'vcs-gitignore-'));
-  });
-
-  it('creates .gitignore with default entries when file does not exist', async () => {
-    await ensureGitignore(tempDir);
-    const content = await readFile(join(tempDir, '.gitignore'), 'utf8');
-    expect(content).toContain('.git/');
-    expect(content).toContain('node_modules/');
-    expect(content).toContain('.jeeves-watcher/');
-    expect(content).toContain('.jeeves-metadata/');
-  });
-
-  it('appends missing entries to existing .gitignore', async () => {
-    await writeFile(
-      join(tempDir, '.gitignore'),
-      '.git/\ncustom-entry\n',
-      'utf8',
-    );
-    await ensureGitignore(tempDir);
-    const content = await readFile(join(tempDir, '.gitignore'), 'utf8');
-    expect(content).toContain('.git/');
-    expect(content).toContain('custom-entry');
-    expect(content).toContain('node_modules/');
-    expect(content).toContain('.jeeves-watcher/');
-    expect(content).toContain('.jeeves-metadata/');
-    // Should not duplicate .git/
-    const gitEntries = content.split('\n').filter((l) => l.trim() === '.git/');
-    expect(gitEntries).toHaveLength(1);
-  });
-
-  it('does nothing when all entries already present', async () => {
-    const existing =
-      '.git/\nnode_modules/\n.jeeves-watcher/\n.jeeves-metadata/\n';
-    await writeFile(join(tempDir, '.gitignore'), existing, 'utf8');
-    await ensureGitignore(tempDir);
-    const content = await readFile(join(tempDir, '.gitignore'), 'utf8');
-    expect(content).toBe(existing);
-  });
-
-  it('handles existing file without trailing newline', async () => {
-    await writeFile(
-      join(tempDir, '.gitignore'),
-      '.git/\nnode_modules/',
-      'utf8',
-    );
-    await ensureGitignore(tempDir);
-    const content = await readFile(join(tempDir, '.gitignore'), 'utf8');
-    expect(content).toContain('.jeeves-watcher/');
-    expect(content).toContain('.jeeves-metadata/');
-    const lines = content.split('\n');
-    expect(lines.filter((l) => l.length > 0).length).toBeGreaterThanOrEqual(4);
-  });
-
-  it('includes custom always-on entries', async () => {
-    await ensureGitignore(tempDir, ['*.log', 'tmp/']);
-    const content = await readFile(join(tempDir, '.gitignore'), 'utf8');
-    expect(content).toContain('*.log');
-    expect(content).toContain('tmp/');
-  });
-});
 
 // ─── Instance methods ───
